Added support for .getAvg() called before .reading()
Previously, calling .getAvg() would result in a CPU crash (divide by 0) if it was called prior to there being any readings yet.  While this is possibly an abnormal use case (checking an average before adding a reading), it does occur in some asynchronous task threads, therefore adding an error-handling save for this scenario is useful.